requirements.txt
torch>=1.9.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0
matplotlib>=3.4.0
seaborn>=0.11.0
Pillow>=8.0.0
tqdm>=4.62.0
pyyaml>=5.4.0
